Check if iptable rule exist only for master nodes

+++ pkg/monitor/dynkeepalived.go
@@ -350,28 +350,31 @@ func KeepalivedWatch(kubeconfigPath, clusterConfigPath, templatePath, cfgPath st
 			}
 			prevConfig = &newConfig
 
-			// Signal to keepalived whether the haproxy firewall rule is in place
-			ruleExists, err := checkHAProxyFirewallRules(apiVip.String(), apiPort, lbPort)
-			if err != nil {
-				log.Error("Failed to check for haproxy firewall rule")
-			} else {
-				_, err := os.Stat(iptablesFilePath)
-				fileExists := !os.IsNotExist(err)
-				if ruleExists {
-					if !fileExists {
-						_, err := os.Create(iptablesFilePath)
-						if err != nil {
-							log.WithFields(logrus.Fields{"path": iptablesFilePath}).Error("Failed to create file")
-						}
-					}
+			if os.Getenv("IS_BOOTSTRAP") == "no" {
+				// Signal to keepalived whether the haproxy firewall rule is in place
+				ruleExists, err := checkHAProxyFirewallRules(apiVip.String(), apiPort, lbPort)
+				if err != nil {
+					log.Error("Failed to check for haproxy firewall rule")
 				} else {
-					if fileExists {
-						err := os.Remove(iptablesFilePath)
-						if err != nil {
-							log.WithFields(logrus.Fields{"path": iptablesFilePath}).Error("Failed to remove file")
+					_, err := os.Stat(iptablesFilePath)
+					fileExists := !os.IsNotExist(err)
+					if ruleExists {
+						if !fileExists {
+							_, err := os.Create(iptablesFilePath)
+							if err != nil {
+								log.WithFields(logrus.Fields{"path": iptablesFilePath}).Error("Failed to create file")
+							}
+						}
+					} else {
+						if fileExists {
+							err := os.Remove(iptablesFilePath)
+							if err != nil {
+								log.WithFields(logrus.Fields{"path": iptablesFilePath}).Error("Failed to remove file")
+							}
 						}
 					}
 				}
+
 			}
 			time.Sleep(interval)
 		}
